Kom mit

+++ ft_memcpy.c
@@ -1,4 +1,18 @@
-void *ft_memcpy(void *dest, const void *src, unsigned int n) {
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memcpy.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jepenoy <jepenoy@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/10/15 11:35:57 by jepenoy           #+#    #+#             */
+/*   Updated: 2024/10/15 11:36:10 by jepenoy          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include<stdlib.h>
+
+void *ft_memcpy(void *dest, const void *src, size_t n) {
     unsigned char *d = (unsigned char *)dest;
     const unsigned char *s = (const unsigned char *)src;
